requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appnope==0.1.4
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.3.0
audioread==3.0.1
bunch==1.0.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
-e git+ssh://git@github.com/Michiel-DK/environmental_sound.git@c68331919d210b50f562d70b51f102258fb3817c#egg=environmental_sound
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.69.0
h5py==3.12.1
hydra-core==1.3.2
idna==3.10
ipdb==0.13.13
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.8.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
lightning-utilities==0.11.9
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==4.25.5
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.1
sentry-sdk==2.20.0
setproctitle==1.3.4
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
soundfile==0.13.0
soxr==0.5.0.post1
stack-data==0.6.3
stringzilla==3.11.3
sympy==1.13.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
termcolor==2.5.0
threadpoolctl==3.5.0
tomli==2.2.1
torch==2.5.1
torchmetrics==1.6.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wandb==0.19.3
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
